3 Shiny Methods, Test data methods, moving queries to folder.
Big update, I was a little behind. We've got 3 new methods specifically aimed at the Shiny app:
/get_map_points
/get_observation_table
/get_observation_details
I've also moved the larger queries into files in a "Queries" folder.
Some changes that are laying groundwork: I've imported pandas and pyarrow and added them to the Dockerfile, and added /get_map_points_parquet. This is mostly for my own testing, but could be useful if performance on the Shiny app starts to be an issue.