Refactor/optimize Plot Observation retrieval and clean up names
This follows directly in line with what we introduced for the Plant
Concept GET endpoints in terms of refactoring and SQL query design,
while also now consistently returning column names in snake_case. It
also adds numerous columns to the individual endpoint that were
previously only in the collection endpoint, and adds a few stratum
detail columns to the collection endpoint that were previously only in
the individual endpoint.