Refactor and optimize Stratum Method retrieval
This follows directly in line with what we introduced for the other GET
endpoints in terms of refactoring and SQL query design. It includes an
extra join in the base query to preserve the original behavior of
returning separate records for different stratum types associated with a
given stratum method. Expectation is that this will be changed in the
future.