Add count reporting to Parquet GET responses
Counts are now added to Parquet k-v metadata with key 'vb_count'.

Previously we reported record counts only in JSON responses, and all of
the logic for getting these counts lived in a method dedicated to
creating the JSON response. Now this logic (including DB queries as
needed) lives in the core response getter, and both the JSON and Parquet
methods only need to know how to take a passed-in count object and
package it up into the response in a useful manner.