Align Taxon Observations with new GET query pattern
This completes the transition of all core VB resource GET endpoints to
use the new query pattern with dynamic SQL generation, rather than
relying on static SQL files. It also simplifies this endpoint to return
either a single record by ID, or the full collection with pagination,
without the extra parameter for (and complexity of) limiting to N
records per plot observation. This was previously implemented for a web
UI requirement that is now met by the plot observation endpoint.