Party upload sql, endpoint, and table defs
Party is our first type in this new format that doesn't have any foreign keys to validate. I don't love having an essentially blank sql file for validate_parties, I might switch this to a flag that decides whether or not it needs to be validated to avoid a null pointer error, that way we don't need a useless file.