Making stem location optional, adding error reporting for the new case that comes up
If a user supplies no stem location fields for a given row, we do not insert that stem location row (note: this doesn't affect stem count, those will still be inserted as normal). If a user supplies a user_sl_code and no other stem location fields, that record will also not be inserted. If a user supplies data in the stem location fields but no user_sl_code, they will see a specific error for that case. In either of the other cases, there is no error, the insert simply does not happen and that stem location record will show nothing in the count/resources sections of the response.